requirements.txt
PyYAML==6.0.1
python-telegram-bot==21.3
smspdudecoder==2.1.1
